+++ Forge/src/main/java/com/christofmeg/justenoughbreeding/rei/BreedingCategoryREI.java
@@ -79,6 +79,10 @@ public class BreedingCategoryREI implements DisplayCategory<BreedingDisplay> {
             } else if (recipe.animalTrusting != null) {
                 Component trusting = Component.translatable("translation.justenoughbreeding.trusting");
                 entityNameString += " (" + trusting.getString() + ")";
+            } else if (recipe.spawnEgg.getDescriptionId().startsWith("item.tfc")) {
+                Component familiarity = Component.translatable("tfc.jade.familiarity");
+                String tfc = familiarity.getString().replaceAll(":[^:]*$", "");
+                entityNameString += " (" + tfc + " > 30" + ")";
             }
 
             int stringWidth = font.width(entityNameString); // Measure the width of the string in pixels