+++ NeoForge/changelog.md
@@ -1,2 +1,4 @@
 1.3.0:
 - Ambient Additions Integration
+- Terra Firma Craft Integration
+- Implement [#15](https://github.com/Christofmeg/JustEnoughBreeding/issues/15)